Feature: Added following features:
The StudentInformation structs are stored as an array (or other suitable data structure) inside a separate model class.
The logout button causes the Student Locations Tabbed View to dismiss, and logs out of the current session.
An activity indicator is displayed during geocoding, and returns to normal state on completion.